Remove TestFilterAvailability because it isn't really useful.
We are linking against the library directly which will answer the question if
the filter is available. It becomes a simple case of if the test compiles
then the filter is available.

Also adjust the input file for the synthetic test file so that it is available
from where ever the test is run.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>